refactor(AWS Deploy): Introduce `getCreateChangeSetParams`
