+++ lib/plugins/aws/deploy/index.js
@@ -7,6 +7,7 @@ const setBucketName = require('../lib/setBucketName');
 const checkForChanges = require('./lib/checkForChanges');
 const monitorStack = require('../lib/monitorStack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
+const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
 const createStack = require('./lib/createStack');
 const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
@@ -44,7 +45,8 @@ class AwsDeploy {
       updateStack,
       monitorStack,
       checkIfBucketExists,
-      waitForChangeSetCreation
+      waitForChangeSetCreation,
+      getCreateChangeSetParams
     );
 
     this.getFileStats = memoize(this.getFileStats.bind(this), { promise: true });
