+++ lib/plugins/aws/deploy/lib/createStack.js
@@ -12,54 +12,15 @@ module.exports = {
     progress.get('main').notice('Creating CloudFormation stack', { isMainEvent: true });
     const stackName = this.provider.naming.getStackName();
     const changeSetName = this.provider.naming.getStackChangeSetName();
-
-    let stackTags = { STAGE: this.provider.getStage() };
-
-    // Merge additional stack tags
-    if (this.serverless.service.provider.stackTags) {
-      const customKeys = Object.keys(this.serverless.service.provider.stackTags);
-      const collisions = Object.keys(stackTags).filter((defaultKey) =>
-        customKeys.some((key) => defaultKey.toLowerCase() === key.toLowerCase())
-      );
-
-      // Delete collisions upfront
-      for (const key of collisions) {
-        delete stackTags[key];
-      }
-
-      stackTags = Object.assign(stackTags, this.serverless.service.provider.stackTags);
-    }
-
-    const createChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-      ChangeSetType: 'CREATE',
-      Capabilities: ['CAPABILITY_IAM', 'CAPABILITY_NAMED_IAM'],
-      Parameters: [],
-      TemplateBody: JSON.stringify(this.serverless.service.provider.coreCloudFormationTemplate),
-      Tags: Object.keys(stackTags).map((key) => ({ Key: key, Value: stackTags[key] })),
-    };
+    const createChangeSetParams = this.getCreateChangeSetParams({
+      changeSetType: 'CREATE',
+      templateBody: this.serverless.service.provider.coreCloudFormationTemplate,
+    });
 
     const executeChangeSetParams = {
       StackName: stackName,
       ChangeSetName: changeSetName,
     };
-
-    if (
-      this.serverless.service.provider.compiledCloudFormationTemplate &&
-      this.serverless.service.provider.compiledCloudFormationTemplate.Transform
-    ) {
-      createChangeSetParams.Capabilities.push('CAPABILITY_AUTO_EXPAND');
-    }
-
-    const customDeploymentRole = this.provider.getCustomDeploymentRole();
-    if (customDeploymentRole) {
-      createChangeSetParams.RoleARN = customDeploymentRole;
-    }
-
-    if (this.serverless.service.provider.notificationArns) {
-      createChangeSetParams.NotificationARNs = this.serverless.service.provider.notificationArns;
-    }
 
     if (this.serverless.service.provider.disableRollback) {
       executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
