+++ lib/plugins/aws/lib/updateStack.js
@@ -14,60 +14,19 @@ module.exports = {
     const stackName = this.provider.naming.getStackName();
     const changeSetName = this.provider.naming.getStackChangeSetName();
 
-    let stackTags = { STAGE: this.provider.getStage() };
     const compiledTemplateFileName = this.provider.naming.getCompiledTemplateS3Suffix();
     const s3Endpoint = getS3EndpointForRegion(this.provider.getRegion());
     const templateUrl = `https://${s3Endpoint}/${this.bucketName}/${this.serverless.service.package.artifactDirectoryName}/${compiledTemplateFileName}`;
 
-    // Merge additional stack tags
-    if (this.serverless.service.provider.stackTags) {
-      const customKeys = Object.keys(this.serverless.service.provider.stackTags);
-      const collisions = Object.keys(stackTags).filter((defaultKey) =>
-        customKeys.some((key) => defaultKey.toLowerCase() === key.toLowerCase())
-      );
-
-      // Delete collisions upfront
-      for (const key of collisions) {
-        delete stackTags[key];
-      }
-
-      stackTags = Object.assign(stackTags, this.serverless.service.provider.stackTags);
-    }
-
-    const createChangeSetParams = {
-      StackName: stackName,
-      ChangeSetName: changeSetName,
-      ChangeSetType: 'CREATE',
-      Capabilities: ['CAPABILITY_IAM', 'CAPABILITY_NAMED_IAM'],
-      Parameters: [],
-      TemplateURL: templateUrl,
-      Tags: Object.keys(stackTags).map((key) => ({ Key: key, Value: stackTags[key] })),
-    };
+    const createChangeSetParams = this.getCreateChangeSetParams({
+      changeSetType: 'CREATE',
+      templateUrl,
+    });
 
     const executeChangeSetParams = {
       StackName: stackName,
       ChangeSetName: changeSetName,
     };
-
-    if (
-      this.serverless.service.provider.compiledCloudFormationTemplate &&
-      this.serverless.service.provider.compiledCloudFormationTemplate.Transform
-    ) {
-      createChangeSetParams.Capabilities.push('CAPABILITY_AUTO_EXPAND');
-    }
-
-    const customDeploymentRole = this.provider.getCustomDeploymentRole();
-    if (customDeploymentRole) {
-      createChangeSetParams.RoleARN = customDeploymentRole;
-    }
-
-    if (this.serverless.service.provider.notificationArns) {
-      createChangeSetParams.NotificationARNs = this.serverless.service.provider.notificationArns;
-    }
-
-    if (this.serverless.service.provider.stackParameters) {
-      createChangeSetParams.Parameters = this.serverless.service.provider.stackParameters;
-    }
 
     if (this.serverless.service.provider.disableRollback) {
       executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
@@ -111,72 +70,16 @@ module.exports = {
     legacy.log('Updating Stack...');
     const stackName = this.provider.naming.getStackName();
     const changeSetName = this.provider.naming.getStackChangeSetName();
-    let stackTags = { STAGE: this.provider.getStage() };
 
-    // Merge additional stack tags
-    if (this.serverless.service.provider.stackTags) {
-      const customKeys = Object.keys(this.serverless.service.provider.stackTags);
-      const collisions = Object.keys(stackTags).filter((defaultKey) =>
-        customKeys.some((key) => defaultKey.toLowerCase() === key.toLowerCase())
-      );
-
-      // Delete collisions upfront
-      for (const key of collisions) {
-        delete stackTags[key];
-      }
-
-      stackTags = Object.assign(stackTags, this.serverless.service.provider.stackTags);
-    }
-
-    const createChangeSetParams = {
-      StackName: stackName,
-      Capabilities: ['CAPABILITY_IAM', 'CAPABILITY_NAMED_IAM'],
-      Parameters: [],
-      TemplateURL: templateUrl,
-      Tags: Object.keys(stackTags).map((key) => ({ Key: key, Value: stackTags[key] })),
-      ChangeSetName: changeSetName,
-      ChangeSetType: 'UPDATE',
-    };
+    const createChangeSetParams = this.getCreateChangeSetParams({
+      changeSetType: 'UPDATE',
+      templateUrl,
+    });
 
     const executeChangeSetParams = {
       StackName: stackName,
       ChangeSetName: changeSetName,
     };
-
-    if (
-      this.serverless.service.provider.compiledCloudFormationTemplate &&
-      this.serverless.service.provider.compiledCloudFormationTemplate.Transform
-    ) {
-      createChangeSetParams.Capabilities.push('CAPABILITY_AUTO_EXPAND');
-    }
-
-    const customDeploymentRole = this.provider.getCustomDeploymentRole();
-    if (customDeploymentRole) {
-      createChangeSetParams.RoleARN = customDeploymentRole;
-    }
-
-    if (this.serverless.service.provider.notificationArns) {
-      createChangeSetParams.NotificationARNs = this.serverless.service.provider.notificationArns;
-    }
-
-    if (this.serverless.service.provider.stackParameters) {
-      createChangeSetParams.Parameters = this.serverless.service.provider.stackParameters;
-    }
-
-    // Policy must have at least one statement, otherwise no updates would be possible at all
-    if (
-      this.serverless.service.provider.stackPolicy &&
-      Object.keys(this.serverless.service.provider.stackPolicy).length
-    ) {
-      createChangeSetParams.StackPolicyBody = JSON.stringify({
-        Statement: this.serverless.service.provider.stackPolicy,
-      });
-    }
-
-    if (this.serverless.service.provider.rollbackConfiguration) {
-      createChangeSetParams.RollbackConfiguration =
-        this.serverless.service.provider.rollbackConfiguration;
-    }
 
     if (this.serverless.service.provider.disableRollback) {
       executeChangeSetParams.DisableRollback = this.serverless.service.provider.disableRollback;
