+++ lib/plugins/aws/rollback.js
@@ -4,6 +4,7 @@ const setBucketName = require('./lib/setBucketName');
 const updateStack = require('./lib/updateStack');
 const monitorStack = require('./lib/monitorStack');
 const waitForChangeSetCreation = require('./lib/wait-for-change-set-creation');
+const getCreateChangeSetParams = require('./lib/get-create-change-set-params');
 const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
 const ServerlessError = require('../../serverless-error');
 const { style, log, progress, legacy } = require('@serverless/utils/log');
@@ -22,7 +23,8 @@ class AwsRollback {
       setBucketName,
       updateStack,
       monitorStack,
-      waitForChangeSetCreation
+      waitForChangeSetCreation,
+      getCreateChangeSetParams
     );
 
     this.hooks = {
